Disable flash memory during deep sleep mode - CS pin is tied to HIGH before entering sleep Move some wifi and system functions to system.h
